Update bitshower.py
This version must be used with python (java to python interpreter). The notes here has been mapped directly from the bits on the screen- the pitches and panning is based on the bit's distance across the screen width, while the volume is based on whether the bit is a 0 or 1.